<예약관리>[Feat] 자산목록 리스트 db연결

+++ IAmGroupee/src/main/java/com/kh/iag/resv/entity/ResvDto.java
@@ -6,10 +6,10 @@ import lombok.Data;
 @Data
 public class ResvDto {
     
-    private long resvNo;
-	private long userNo;
-	private long roomNo;
-	private long equipNo;
+    private int resvNo;
+	private int userNo;
+	private int roomNo;
+	private int equipNo;
 	private Date resvStart;
 	private Date resvEnd;
 	private Date createDate;
